Combining martech.js into utils.js to reduce scripts in critical path leading to interact call.
